greybus: loopback: use u32 for stats update
The only values supplied to gb_loopback_update_stats() are 32-bits,
so change the type of the second argument to reflect that.

+++ drivers/staging/greybus/loopback.c
@@ -348,7 +348,7 @@ static void gb_loopback_reset_stats(struct gb_loopback *gb)
 	       sizeof(struct gb_loopback_stats));
 }
 
-static void gb_loopback_update_stats(struct gb_loopback_stats *stats, u64 val)
+static void gb_loopback_update_stats(struct gb_loopback_stats *stats, u32 val)
 {
 	if (stats->min > val)
 		stats->min = val;
